Refactor AddressBook to StudentBook and Person to Student in test
