+++ src/test/java/jarvis/testutil/StudentUtil.java
@@ -12,9 +12,9 @@ import jarvis.model.Student;
 public class StudentUtil {
 
     /**
-     * Returns an add command string for adding the {@code student}.
+     * Returns an add student command string for adding the {@code student}.
      */
-    public static String getAddCommand(Student student) {
+    public static String getAddStudentCommand(Student student) {
         return AddStudentCommand.COMMAND_WORD + " " + getStudentDetails(student);
     }
