+++ src/test/java/jarvis/testutil/TestUtil.java
@@ -8,6 +8,7 @@ import java.nio.file.Paths;
 import jarvis.commons.core.index.Index;
 import jarvis.model.Model;
 import jarvis.model.Student;
+import jarvis.model.Task;
 
 /**
  * A utility class for test cases.
@@ -33,23 +34,30 @@ public class TestUtil {
     }
 
     /**
-     * Returns the middle index of the person in the {@code model}'s person list.
+     * Returns the middle index of the student in the {@code model}'s student list.
      */
     public static Index getMidIndex(Model model) {
         return Index.fromOneBased(model.getFilteredStudentList().size() / 2);
     }
 
     /**
-     * Returns the last index of the person in the {@code model}'s person list.
+     * Returns the last index of the student in the {@code model}'s student list.
      */
     public static Index getLastIndex(Model model) {
         return Index.fromOneBased(model.getFilteredStudentList().size());
     }
 
     /**
-     * Returns the person in the {@code model}'s person list at {@code index}.
+     * Returns the student in the {@code model}'s student list at {@code index}.
      */
-    public static Student getPerson(Model model, Index index) {
+    public static Student getStudent(Model model, Index index) {
         return model.getFilteredStudentList().get(index.getZeroBased());
+    }
+
+    /**
+     * Returns the task in the {@code model}'s task list at {@code index}.
+     */
+    public static Task getTask(Model model, Index index) {
+        return model.getFilteredTaskList().get(index.getZeroBased());
     }
 }
